Update blacklist.sh
Updated the script to put the blacklist creation steps into a function so that an if, then, elif, then, else statement could be created.  This new series of checks determines if the blacklist backup file exists on a reboot.  If it does, it restores that backup list to speed up provisioning.  If it does not then it creates a new list on reboot to ensure the device is protected rather than waiting until the next scheduled interval to process and create the list.  The list will still be processed and updated at the scheduled interval as well.  Additional logging is included to state whether the list is being restored from a backup or processed as new.