Update blacklist.sh
Added additional code to do a list comparison and write the results to the log file.  This will show changes between each run of the script on what IPs were added or removed from the new list created by the script.  This comparison will not run on a reboot of the USG to reduce provisioning time.  Also changed "FireHOL" back to "Dynamic Threat List" to make sure @brontide instructions continue to work with this script.