Rename file
Renaming file from blacklist.sh to blocklist.sh.  This update also includes all references inside the file to blacklist are now updated to blocklist.  Additional logging directed to console by using echo so if script is run manually you have a better idea of what step it is on and what it is doing.